Adding render ability.
Adding the ability to render XML from the structures in the parser. This
is primarally intended for read-modify-write cycles. There are also a
series of tests to ensure that modifications don't break the rendering.
It doesn't format the output nor ensure that you have valid XML characters
in your strings.