Attributes should be a map.
Fixed a spelling error, change attribute to be std::pair instead of
custom class. Change attributes to be a map instead of a list. This
causes us to loose the ordering, but according to the XML spec, it's
not important. We will be consistent though, they will be rendered
in alphabetical order.