Switched over to StringRef instead of std::string
This allows us to reference the file the strings are coming from
rather than holding the entire string in memory. Currently they are 24
bytes in length, so for every string we store that is less than 24
bytes, we are actually making the memory situation worse. I think this
can be improved with some tweaking.