Switch to StringT
This switches over the string type to StringT. This is a polymorphic
string type that can store either a string literal (StringLit) or a
string reference (StringRef). The string reference type references a
location and length in a file. This allows a developer to save some RAM
when dealing with XML files that have large strings embedded in them.

This also includes a bug fix for FileString that puts an IO mutex in
place in the substr function. This will prevent multiple threads from
trying to do IO at the same time.